feat: centralize dashboard layout
