--- app/science/page.tsx
@@ -1,29 +1,0 @@
-import SidebarNav from "@/components/SidebarNav"
-import { TempHumidityChart, VPDGauge } from "@/components/Charts"
-import Footer from "@/components/Footer"
-
-export default function SciencePanel() {
-  return (
-    <div className="flex">
-      <SidebarNav />
-      <main className="flex-1 p-6">
-        <header className="backdrop-blur bg-white/80 sticky top-0 z-10 p-4 flex items-center justify-between shadow-sm">
-          <h2 className="font-semibold text-xl">Science Panel</h2>
-          <button className="px-3 py-1 border rounded hover:bg-gray-100">°F / °C</button>
-        </header>
-
-        <section className="mt-6">
-          <h3 className="font-medium text-gray-800">Environment Data</h3>
-          <TempHumidityChart />
-        </section>
-
-        <section className="mt-6">
-          <h3 className="font-medium text-gray-800">VPD Gauge</h3>
-          <VPDGauge />
-        </section>
-
-        <Footer />
-      </main>
-    </div>
-  )
-}
